ENH: Sort subclasses for cross-platform consistency (#128)
* Sort subclasses for cross-platform consistency

The sort order for subclasses is inconsistent between platforms. As a
result, projects that include generated docs in their repos see diffs to
the generated files based on the platform of the contributor. This adds
an explicit sort so we have a deterministic order of subclasses.

* Change intrinsic sort order to use refname

This will sort classes using their fully qualified names.