feat: Also scan modules with `inspect.ismodule` on public objects.
When documenting an extension, (e.g. a `.so` file), `iter_modules`
can't be used to scan submodules because there is no files or
directories to iterate over. Nonetheless, the `<package>.__all__`
contains public objects of kind `ModuleType`. Simply by collecting
submodules to `Module.doc` with `inspect.ismodule` makes it possible
to generate documentation for submodules of an extension.